Major: Added RoPE backward operation, RoPE backward has been embedded in llama attention layer Test for W_Q grad is working, test for W_K grad is NOT.
Minor: little fix in rope tensor level, to ensure it will work both with 4
and 5 dimensional tensors.

+++ wrappers/python/tests/layer/test_llama_attention.py
@@ -75,7 +75,7 @@ TEST_PARAMS = [
             n_head=16,
             n_head_tile=8,
             n_head_kv=4,
-            dtype='bf16',
+            dtype='fp32',
             bias=False,
         ),
         # marks=[
@@ -192,7 +192,7 @@ def generate_inputs(params: LlamaAttentionTestParams):
             size=(params.n_batch, params.n_seq),
             dtype=np.int64
     )
-    #pos_ids = np.zeros((params.n_batch, params.n_seq), dtype=np.int64)
+    # pos_ids = np.zeros((params.n_batch, params.n_seq), dtype=np.int64)
     pos_ids_torch = torch.tensor(pos_ids, dtype=torch.long)
 
     nntile_layer = nntile.layer.LlamaAttention.from_torch(
